phubme v0.5.2 PhubMe.Slack

Send message to slack. With nicknames, it takes the %IssueComment struct :

It sends a message to each nicknames then log “All procceed”

Without nicknames :

It log “All procceed”

When error :

It log the error message All logs have specific level that can be defined in mix config per env.

Functions

send_private_message(issue_comment)